Moved constants to separate file in CRUD
