+++ lib/services/crud/notes_service.dart
@@ -7,6 +7,7 @@ import 'package:path_provider/path_provider.dart'
     show MissingPlatformDirectoryException, getApplicationDocumentsDirectory;
 import 'package:path/path.dart' show join;
 
+import 'crud_constants.dart';
 import 'crud_exceptions.dart';
 
 class NotesService {
@@ -267,33 +268,3 @@ class DatabaseNote {
   @override
   int get hashCode => id.hashCode;
 }
-
-const dbName = 'notes.db';
-const userTable = 'user';
-const noteTable = 'note';
-
-const idColumn = 'id';
-const emailColumn = 'email';
-const userIdColumn = 'user_id';
-const textColumn = 'text';
-const isSyncedColumn = 'is_synced_with_cloud';
-
-/// Create the tables
-const createUserTable = '''
-      CREATE TABLE IF NOT EXISTS 'user' (
-        'id' INTEGER NOT NULL,
-        'email' TEXT NOT NULL UNIQUE,
-        PRIMARY KEY('id' AUTOINCREMENT)
-      );
-      ''';
-
-const createNoteTable = '''
-        CREATE TABLE IF NOT EXISTS 'note' (
-        'id' INTEGER NOT NULL,
-        'user_id' INTEGER NOT NULL,
-        'text' TEXT,
-        'is_synced_with_cloud' INTEGER NOT NULL DEFAULT 0,
-        FOREIGN KEY('user_id') REFERENCES 'user'('id'),
-        PRIMARY KEY('id' AUTOINCREMENT)
-      );
-      ''';
